Refactored heavy code out of system_manager
